sdf: restructure code
the first condition will meat for voxels that have been successfully
calculated. Which gives them a early return, if a complete local work
group is done, it is faster completly done.

After that we are only calculcating the neightbour_distance if we are
not knowing the neightbour distance yet. And conditionally write a new
one.

In the end we are writing.